move kms client to devenv
