+++ deployment/environment/devenv/internal/kms/kms_client.go
@@ -1,4 +1,4 @@
-package deployment
+package kms
 
 import (
 	"bytes"
