+++ deployment/environment/devenv/internal/kms/kms_client_test.go
@@ -1,4 +1,4 @@
-package deployment
+package kms
 
 import (
 	"encoding/hex"
